requirements.txt
dida365>=0.1.6
langchain>=0.3.14
langchain-openai>=0.3.1
python-dotenv>=1.0.1
openai>=1.59.8
gradio>=5.12.0 